fix: remove all unused components
