--- packages/user-interface/src/components/Layout.module.scss
@@ -1,53 +1,0 @@
-@use "../styles/mixins";
-
-.page-container {
-  width: auto;
-  display: flex;
-  justify-content: center;
-  box-sizing: border-box;
-  min-height: 100%;
-  margin-inline-start: var(--nlportal-space-inline-xl, 1.5rem);
-  margin-inline-end: var(--nlportal-space-inline-xl, 1.5rem);
-  padding-block-end: var(--nlportal-space-block-3xl, 2.5rem);
-
-  @include mixins.breakpoint(tablet) {
-    padding-block-end: var(--nlportal-space-block-5xl, 4rem);
-  }
-}
-
-.page-container__inner {
-  width: 100%;
-  max-width: var(--nlportal-max-page-width, 1024px);
-  display: flex;
-  justify-content: space-between;
-  padding-block-start: var(--nlportal-space-block-3xl, 2.5rem);
-
-  @include mixins.breakpoint(tablet) {
-    padding-block-start: var(--nlportal-space-block-5xl, 4rem);
-  }
-}
-
-.page-container__menu {
-  width: 0;
-  z-index: var(--nlportal-z-index-menu, 200);
-
-  @include mixins.breakpoint(tablet) {
-    width: var(--nlportal-desktop-menu-width, 230px);
-    position: fixed;
-  }
-}
-
-.page-container__page {
-  width: 100%;
-
-  @include mixins.breakpoint(tablet) {
-    margin-left: calc(
-      var(--nlportal-desktop-menu-width, 230px) +
-        var(--nlportal-space-inline-2xl, 2rem)
-    );
-  }
-}
-
-.page-container__page--fullscreen {
-  margin-left: 0;
-}
