--- packages/user-interface/src/components/MetaIcon.module.scss
@@ -1,49 +1,0 @@
-@use "../styles/mixins";
-
-.meta-icon {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  box-sizing: border-box;
-  position: relative;
-  width: 100%;
-  margin-block-start: var(--nlportal-space-block-3xl, 2.5rem);
-  margin-block-end: var(--nlportal-space-block-3xl, 2.5rem);
-  padding-inline-start: var(--nlportal-space-block-md, 1rem);
-  padding-inline-end: var(--nlportal-space-block-md, 1rem);
-
-  @include mixins.breakpoint(mobile) {
-    width: 50%;
-  }
-
-  @include mixins.breakpoint(desktop) {
-    width: 25%;
-  }
-}
-
-.meta-icon__icon {
-  font-size: var(--nlportal-meta-icon-size, 2rem) !important;
-  color: var(--nlportal-meta-icon-color, hsl(207 80% 35%));
-  margin-block-end: var(--nlportal-space-block-xs, 0.5rem);
-}
-
-.meta-icon__paragraph {
-  line-height: var(--nlportal-meta-icon-line-height, 1.5rem) !important;
-  max-width: 100%;
-  text-align: center;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.meta-icon__paragraph--bold {
-  font-weight: var(--nlportal-meta-icon-font-weight-bold, 700) !important;
-}
-
-.meta-icon__border {
-  background-color: var(--nlportal-meta-icon-border-color, hsl(0 0% 82%));
-  height: 50%;
-  width: 1px;
-  position: absolute;
-  right: 0;
-  top: 25%;
-}
