--- packages/user-interface/src/components/MetaIcon.tsx
@@ -1,48 +1,0 @@
-import * as React from "react";
-import { ReactElement } from "react";
-import { Paragraph } from "@gemeente-denhaag/typography";
-import Skeleton from "react-loading-skeleton";
-import classNames from "classnames";
-import { useIntl } from "react-intl";
-import styles from "./MetaIcon.module.scss";
-
-interface MetaIconProps {
-  title: string;
-  subtitle?: string | ReactElement;
-  icon: ReactElement<HTMLDivElement>;
-  showRightBorder?: boolean;
-}
-
-const MetaIcon = ({
-  title,
-  subtitle,
-  icon,
-  showRightBorder,
-}: MetaIconProps) => {
-  const intl = useIntl();
-
-  return (
-    <div className={styles["meta-icon"]}>
-      {React.cloneElement(icon, { className: styles["meta-icon__icon"] })}
-      <Paragraph className={styles["meta-icon__paragraph"]}>{title}</Paragraph>
-      <Paragraph
-        className={classNames(
-          styles["meta-icon__paragraph"],
-          styles["meta-icon__paragraph--bold"],
-        )}
-      >
-        {subtitle || (
-          <span
-            aria-busy
-            aria-disabled
-            aria-label={intl.formatMessage({ id: "element.loading" })}
-          >
-            <Skeleton width={100} />
-          </span>
-        )}
-      </Paragraph>
-      {showRightBorder && <div className={styles["meta-icon__border"]} />}
-    </div>
-  );
-};
-export default MetaIcon;
