--- packages/user-interface/src/components/MobileMenuButton.module.scss
@@ -1,10 +1,0 @@
-.denhaag-button--mobile-menu {
-  height: var(--nlportal-mobile-menu-element-height, 43px);
-  color: var(--nlportal-mobile-menu-font-color, hsl(138 58% 33%)) !important;
-  background-color: var(
-    --nlportal-mobile-menu-background-color,
-    hsl(0 0% 95%)
-  ) !important;
-  font-weight: var(--nlportal-mobile-menu-font-weight, 700) !important;
-  justify-content: center;
-}
