--- packages/user-interface/src/components/UserName.module.scss
@@ -1,32 +1,0 @@
-@use "../styles/mixins";
-
-.user-name {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  justify-content: center;
-}
-
-.user-name--mobile-menu {
-  width: 100%;
-  background-color: var(--nlportal-mobile-menu-background-color, hsl(0 0% 95%));
-  height: var(--nlportal-mobile-menu-element-height, 43px);
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-}
-
-.user-name__paragraph {
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  max-width: calc(100% - var(--nlportal-space-inline-4xl, 3rem));
-
-  @include mixins.breakpoint(tablet) {
-    max-width: var(--nlportal-header-user-name-max-width-tablet, 275px);
-  }
-
-  @include mixins.breakpoint(desktop) {
-    max-width: var(--nlportal-header-user-name-max-width-desktop, 350px);
-  }
-}
