--- packages/user-interface/src/constants/breakpoints.ts
@@ -1,5 +1,0 @@
-export const BREAKPOINTS = {
-  MOBILE: "(min-width: 320px)",
-  TABLET: "(min-width: 768px)",
-  DESKTOP: "(min-width: 1024px)",
-};
