--- packages/user-interface/src/interfaces/portal-footer.ts
@@ -1,4 +1,0 @@
-export type PortalFooter = Array<{
-  titleTranslationKey: string;
-  links: Array<{ linkTranslationKey: string; url: string; hrefLang: string }>;
-}>;
