--- packages/user-interface/src/tests/components/CurrentPageIndicator.test.tsx
@@ -1,8 +1,0 @@
-import { describe, it, expect } from "vitest";
-import CurrentPageIndicator from "../../components/CurrentPageIndicator";
-
-describe("CurrentPage", () => {
-  it("is truthy", () => {
-    expect(CurrentPageIndicator).toBeTruthy();
-  });
-});
