--- packages/user-interface/src/tests/components/LanguageSwitcher.test.tsx
@@ -1,8 +1,0 @@
-import { describe, it, expect } from "vitest";
-import LanguageSwitcher from "../../components/LanguageSwitcher";
-
-describe("LanguageSwitcher", () => {
-  it("is truthy", () => {
-    expect(LanguageSwitcher).toBeTruthy();
-  });
-});
