--- packages/user-interface/src/tests/components/ListItemSkeleton.test.tsx
@@ -1,8 +1,0 @@
-import { describe, it, expect } from "vitest";
-import ListItemSkeleton from "../../components/ListItemSkeleton";
-
-describe("ListItemSkeleton", () => {
-  it("is truthy", () => {
-    expect(ListItemSkeleton).toBeTruthy();
-  });
-});
