--- packages/user-interface/src/tests/components/Logout.test.tsx
@@ -1,8 +1,0 @@
-import { describe, it, expect } from "vitest";
-import Logout from "../../components/Logout";
-
-describe("Logout", () => {
-  it("is truthy", () => {
-    expect(Logout).toBeTruthy();
-  });
-});
