--- packages/user-interface/src/tests/components/MenuItem.test.tsx
@@ -1,8 +1,0 @@
-import { describe, it, expect } from "vitest";
-import MenuItem from "../../components/MenuItem";
-
-describe("MenuItem", () => {
-  it("is truthy", () => {
-    expect(MenuItem).toBeTruthy();
-  });
-});
