--- packages/user-interface/src/tests/components/MenuToggleButton.test.tsx
@@ -1,8 +1,0 @@
-import { describe, it, expect } from "vitest";
-import MenuToggleButton from "../../components/MenuToggleButton";
-
-describe("MenuToggleButton", () => {
-  it("is truthy", () => {
-    expect(MenuToggleButton).toBeTruthy();
-  });
-});
