--- packages/user-interface/src/tests/components/MetaIcon.test.tsx
@@ -1,8 +1,0 @@
-import { describe, it, expect } from "vitest";
-import MetaIcon from "../../components/MetaIcon";
-
-describe("MetaIcon", () => {
-  it("is truthy", () => {
-    expect(MetaIcon).toBeTruthy();
-  });
-});
