--- packages/user-interface/src/tests/components/MobileMenuButton.test.tsx
@@ -1,8 +1,0 @@
-import { describe, it, expect } from "vitest";
-import MobileMenuButton from "../../components/MobileMenuButton";
-
-describe("MobileMenuButton", () => {
-  it("is truthy", () => {
-    expect(MobileMenuButton).toBeTruthy();
-  });
-});
