--- packages/user-interface/src/tests/components/UserName.test.tsx
@@ -1,8 +1,0 @@
-import { describe, it, expect } from "vitest";
-import UserName from "../../components/UserName";
-
-describe("UserName", () => {
-  it("is truthy", () => {
-    expect(UserName).toBeTruthy();
-  });
-});
