--- packages/user-interface/src/tests/providers/LayoutProvider.test.tsx
@@ -1,8 +1,0 @@
-import { describe, it, expect } from "vitest";
-import { LayoutProvider } from "../../contexts/LayoutContext";
-
-describe("LayoutProvider", () => {
-  it("is truthy", () => {
-    expect(LayoutProvider).toBeTruthy();
-  });
-});
